+++ adminsortable2/static/adminsortable2/js/inline-sortable.js
@@ -1,7 +1,9 @@
 // make tabular- and stacked inlines sortable
-jQuery(function($) {
-	$('div.inline-group.sortable').each(function() {
-		var default_order_field = $(this).nextUntil('div.default_order_field').next().attr('default_order_field');
+django.jQuery(function($) {
+	$('div.inline-group.sortable').each(function () {
+		var $order_div = $(this).nextUntil('div.default_order_field').next()
+		var default_order_field = $order_div.attr('default_order_field');
+		var default_order_direction = $order_div.attr('default_order_direction');
 		var order_input_field = 'input[name$="-' + default_order_field + '"]';
 		// first, try with tabluar inlines
 		var tabular_inlines = $(this).find('div.tabular table');
@@ -12,12 +14,16 @@ jQuery(function($) {
 			scroll: true,
 			cursor: 'ns-resize',
 			containment: $(this).find('tbody'),
-			stop: function(event, dragged_rows) {
+			stop: function (event, dragged_rows) {
 				var $result_list = $(this);
-				$result_list.find('tbody tr').each(function(index) {
+				$result_list.find('tbody tr').each(function (index) {
 					$(this).removeClass('row1 row2').addClass(index % 2 ? 'row2' : 'row1');
 				});
-				$result_list.find('tbody tr.has_original').each(function(index) {
+				var originals = $result_list.find('tbody tr.has_original').get()
+				if (default_order_direction === '-1') {
+					originals.reverse();
+				}
+				$(originals).each(function (index) {
 					$(this).find(order_input_field).val(index + 1);
 				});
 			}
@@ -32,9 +38,13 @@ jQuery(function($) {
 			scroll: true,
 			cursor: 'ns-resize',
 			containment: $(this),
-			stop: function(event, dragged_rows) {
+			stop: function (event, dragged_rows) {
 				var $result_list = $(this);
-				$result_list.find('div.inline-related.has_original').each(function(index) {
+				var originals = $result_list.find('div.inline-related.has_original').get()
+				if (default_order_direction === '-1') {
+					originals.reverse();
+				}
+				$(originals).each(function (index) {
 					$(this).find(order_input_field).val(index + 1);
 				});
 			}